Moved back everything into one file
